refactor: DB와 연결하는 역할을 ConnectionManager에게 담당하도록 리팩토링
ConnectionManager DB와 연결을 담당하는 객체
LogDao DB와 연결을 담당하는 부분을 제외하고 ConnectionManager를 이용하도록 수정
