+++ src/main/java/chess/domain/web/ConnectionManager.java
@@ -1,0 +1,35 @@
+package chess.domain.web;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+public class ConnectionManager {
+	public Connection getConnection() {
+		Connection con = null;
+		String server = "localhost:13306"; // MySQL 서버 주소
+		String database = "db_name"; // MySQL DATABASE 이름
+		String option = "?useSSL=false&serverTimezone=UTC";
+		String userName = "root"; //  MySQL 서버 아이디
+		String password = "root"; // MySQL 서버 비밀번호
+
+		// 드라이버 로딩
+		try {
+			Class.forName("com.mysql.cj.jdbc.Driver");
+		} catch (ClassNotFoundException e) {
+			System.err.println(" !! JDBC Driver load 오류: " + e.getMessage());
+			e.printStackTrace();
+		}
+
+		// 드라이버 연결
+		try {
+			con = DriverManager.getConnection("jdbc:mysql://" + server + "/" + database + option, userName, password);
+			System.out.println("정상적으로 연결되었습니다.");
+		} catch (SQLException e) {
+			System.err.println("연결 오류:" + e.getMessage());
+			e.printStackTrace();
+		}
+
+		return con;
+	}
+}
