+++ src/main/java/chess/domain/web/LogDao.java
@@ -1,42 +1,18 @@
 package chess.domain.web;
 
-import java.sql.*;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
 public class LogDao {
-	public Connection getConnection() {
-		Connection con = null;
-		String server = "localhost:13306"; // MySQL 서버 주소
-		String database = "db_name"; // MySQL DATABASE 이름
-		String option = "?useSSL=false&serverTimezone=UTC";
-		String userName = "root"; //  MySQL 서버 아이디
-		String password = "root"; // MySQL 서버 비밀번호
-
-		// 드라이버 로딩
-		try {
-			Class.forName("com.mysql.cj.jdbc.Driver");
-		} catch (ClassNotFoundException e) {
-			System.err.println(" !! JDBC Driver load 오류: " + e.getMessage());
-			e.printStackTrace();
-		}
-
-		// 드라이버 연결
-		try {
-			con = DriverManager.getConnection("jdbc:mysql://" + server + "/" + database + option, userName, password);
-			System.out.println("정상적으로 연결되었습니다.");
-		} catch (SQLException e) {
-			System.err.println("연결 오류:" + e.getMessage());
-			e.printStackTrace();
-		}
-
-		return con;
-	}
+	private final ConnectionManager connectionManager = new ConnectionManager();
 
 	public void insert(String start, String end) throws SQLException {
 		String query = "INSERT INTO log (start, end) VALUES (?, ?)";
-		PreparedStatement pstmt = getConnection().prepareStatement(query);
+		PreparedStatement pstmt = connectionManager.getConnection().prepareStatement(query);
 		pstmt.setString(1, start);
 		pstmt.setString(2, end);
 		pstmt.executeUpdate();
@@ -44,7 +20,7 @@ public class LogDao {
 
 	public Map<Integer, Log> selectAll() throws SQLException {
 		String query = "SELECT * FROM log";
-		PreparedStatement pstmt = getConnection().prepareStatement(query);
+		PreparedStatement pstmt = connectionManager.getConnection().prepareStatement(query);
 		ResultSet rs = pstmt.executeQuery();
 		Map<Integer, Log> result = new LinkedHashMap<>();
 		while (rs.next()) {
@@ -55,10 +31,10 @@ public class LogDao {
 
 	public void clear() throws SQLException {
 		String query = "DELETE FROM log";
-		PreparedStatement pstmt = getConnection().prepareStatement(query);
+		PreparedStatement pstmt = connectionManager.getConnection().prepareStatement(query);
 		pstmt.executeUpdate();
 		query = "ALTER TABLE log AUTO_INCREMENT=1";
-		pstmt = getConnection().prepareStatement(query);
+		pstmt = connectionManager.getConnection().prepareStatement(query);
 		pstmt.executeUpdate();
 	}
 }
